refactors and update

+++ Src/Core/Application/GroupShares/Commands/DeleteGroupShare/DeleteGroupShareCommand.cs
@@ -1,56 +1,51 @@
-﻿using Application.Common.Interfaces.Database;
+﻿using Application.Common.Exceptions;
+using Application.Common.Interfaces.Database;
 using MediatR;
 
-namespace Application.GroupShares.Commands.DeleteGroupShare
+namespace Application.GroupShares.Commands.DeleteGroupShare;
+
+public class DeleteGroupShareCommand : IRequest
 {
     /// <summary>
-    /// Class DeleteGroupShareCommand
+    /// GroupId associated with share
     /// </summary>
-    public class DeleteGroupShareCommand : IRequest
+    public string GroupId { get; set; }
+    /// <summary>
+    /// Shared filename
+    /// </summary>
+    public string Filename { get; set; }
+    
+    public class DeleteGroupShareCommandHandler : IRequestHandler<DeleteGroupShareCommand>
     {
-        /// <summary>
-        /// GroupId associated with share
-        /// </summary>
-        public string GroupId { get; set; }
-        /// <summary>
-        /// Shared filename
-        /// </summary>
-        public string Filename { get; set; }
-
-        /// <summary>
-        /// Class DeleteGroupShareCommandHandler
-        /// </summary>
-        public class DeleteGroupShareCommandHandler : IRequestHandler<DeleteGroupShareCommand>
+        private readonly IUnitOfWork _unitOfWork;
+        
+        public DeleteGroupShareCommandHandler(IUnitOfWork unitOfWork)
         {
-            private readonly IUnitOfWork _unitOfWork;
+            _unitOfWork = unitOfWork;
+        }
 
-            /// <summary>
-            /// Initializes new instance of <see cref="DeleteGroupShareCommandHandler" /> class.
-            /// </summary>
-            /// <param name="unitOfWork">The <see cref="IUnitOfWork"/></param>
-            public DeleteGroupShareCommandHandler(IUnitOfWork unitOfWork)
+        /// <summary>
+        /// Deletes group share
+        /// </summary>
+        /// <param name="request">GroupId and filename</param>
+        /// <param name="cancellationToken">
+        /// <see cref="CancellationToken"/>
+        /// </param>
+        /// <returns>Unit</returns>
+        public async Task<Unit> Handle(DeleteGroupShareCommand request, CancellationToken cancellationToken)
+        {
+            var groupShare = await _unitOfWork.GroupShareRepository.GetObjectBy(x => x.GroupId == Guid.Parse(request.GroupId)
+                && x.Filename == request.Filename, cancellationToken: cancellationToken);
+
+            if (groupShare == null)
             {
-                _unitOfWork = unitOfWork;
+                throw new ShareNotFoundException(request.Filename, request.GroupId);
             }
 
-            /// <summary>
-            /// Deletes group share
-            /// </summary>
-            /// <param name="request">GroupId and filename</param>
-            /// <param name="cancellationToken">
-            /// <see cref="CancellationToken"/>
-            /// </param>
-            /// <returns>Unit</returns>
-            public async Task<Unit> Handle(DeleteGroupShareCommand request, CancellationToken cancellationToken)
-            {
-                var groupShare = await _unitOfWork.GroupShareRepository.GetObjectBy(x => x.GroupId == Guid.Parse(request.GroupId)
-                                                                                         && x.Filename == request.Filename, cancellationToken: cancellationToken);
+            await _unitOfWork.GroupShareRepository.Delete(groupShare);
+            await _unitOfWork.Save(cancellationToken);
 
-                await _unitOfWork.GroupShareRepository.Delete(groupShare);
-                await _unitOfWork.Save(cancellationToken);
-
-                return Unit.Value;
-            }
+            return Unit.Value;
         }
     }
 }
