+++ Src/Core/Application/Groups/Commands/DeleteGroup/DeleteGroupCommand.cs
@@ -1,52 +1,47 @@
-﻿using Application.Common.Interfaces.Database;
+﻿using Application.Common.Exceptions;
+using Application.Common.Interfaces.Database;
 using Domain.Entities;
 using MediatR;
 
-namespace Application.Groups.Commands.DeleteGroup
+namespace Application.Groups.Commands.DeleteGroup;
+
+public class DeleteGroupCommand : IRequest
 {
     /// <summary>
-    /// Class DeleteGroupCommand
+    /// GroupId
     /// </summary>
-    public class DeleteGroupCommand : IRequest
+    public Guid GroupId { get; set; }
+
+    public class DeleteGroupCommandHandler : IRequestHandler<DeleteGroupCommand>
     {
-        /// <summary>
-        /// GroupId
-        /// </summary>
-        public string GroupId { get; set; }
+        private IUnitOfWork _unitOfWork;
 
-        /// <summary>
-        /// Class DeleteGroupCommandHandler
-        /// </summary>
-        public class DeleteGroupCommandHandler : IRequestHandler<DeleteGroupCommand>
+        public DeleteGroupCommandHandler(IUnitOfWork unitOfWork)
         {
-            private IUnitOfWork _unitOfWork;
+            _unitOfWork = unitOfWork;
+        }
 
-            /// <summary>
-            /// Initializes new instance of <see cref="DeleteGroupCommandHandler" /> class.
-            /// </summary>
-            /// <param name="unitOfWork">The <see cref="IUnitOfWork"/></param>
-            public DeleteGroupCommandHandler(IUnitOfWork unitOfWork)
+        /// <summary>
+        /// Deletes group
+        /// </summary>
+        /// <param name="request">GroupId</param>
+        /// <param name="cancellationToken">
+        /// <see cref="CancellationToken"/>
+        /// </param>
+        /// <returns>Unit</returns>
+        public async Task<Unit> Handle(DeleteGroupCommand request, CancellationToken cancellationToken)
+        {
+            var group = await _unitOfWork.GroupRepository.GetObjectBy(x => x.Id == request.GroupId, cancellationToken: cancellationToken);
+
+            if (group is null)
             {
-                _unitOfWork = unitOfWork;
+                throw new GroupNotFoundException(request.GroupId);
             }
-
-            /// <summary>
-            /// Deletes group
-            /// </summary>
-            /// <param name="request">GroupId</param>
-            /// <param name="cancellationToken">
-            /// <see cref="CancellationToken"/>
-            /// </param>
-            /// <returns>Unit</returns>
-            public async Task<Unit> Handle(DeleteGroupCommand request, CancellationToken cancellationToken)
-            {
-                var group = await _unitOfWork.GroupRepository.GetObjectBy(x => x.Id == Guid.Parse(request.GroupId), cancellationToken: cancellationToken);
                 
-                await _unitOfWork.GroupRepository.Delete(group);
-                await _unitOfWork.Save(cancellationToken);
+            await _unitOfWork.GroupRepository.Delete(group);
+            await _unitOfWork.Save(cancellationToken);
 
-                return Unit.Value;
-            }
+            return Unit.Value;
         }
     }
 }
